TEST_V1

+++ delete.php
@@ -1,8 +1,11 @@
 <?php
-require "config.php";
+require_once __DIR__ . '/db_connect.php';
 
-$id = $_GET["id"];
+if(!isset($_GET["PrdID"])){
+    die("ID missing");
+}
 
+$id = $_GET["PrdID"];
 // TODO: Delete
 mysqli_query($conn, "DELETE FROM users WHERE id = $id");
